Unfinished ProbThreshPhenomena

+++ src/main/java/catworks/phenomena/ProbabilityMatrix.java
@@ -3,15 +3,15 @@ package catworks.phenomena;
 import java.lang.System;
 
 public class ProbabilityMatrix {
-  private float [] sizes; //this contains the sizes for each network in the IDN
+  private int [] sizes; //this contains the sizes for each network in the IDN
   private float [][] probs; //this contains the probability of spread for each "section" of the IDN
 
-  public ProbabilityMatrix(float [] sizes, float [][] probs){
+  public ProbabilityMatrix(int [] sizes, float [][] probs){
     this.sizes = sizes;
     this.probs = probs;
   }
 
-  public float [] getSizes(){
+  public int [] getSizes(){
     return sizes;
   }
